correct all trafficviolation endpoints

+++ src/services/GamaCore/Gama.Application/Contracts/TrafficFineManagement/ITrafficViolationService.cs
@@ -5,9 +5,9 @@ namespace Gama.Application.Contracts.TrafficFineManagement;
 
 public interface ITrafficViolationService
 {
-    Task<Result<TrafficViolation>> GetAsync(short id);
+    Task<Result<TrafficViolation>> GetAsync(int id);
     Task<Result<IEnumerable<TrafficViolation>>> GetTrafficsViolationsAsync();
     Task<Result<TrafficViolation>> CreateAsync(TrafficViolation trafficViolation);
     Task<Result<TrafficViolation>> UpdateAsync(TrafficViolation trafficViolation);
-    Task<Result<bool>> DeleteAsync(short id);
+    Task<Result<bool>> DeleteAsync(int id);
 }
